Move TVListViewHandler into sample project
