+++ sample/Sample/MauiProgram.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Maui.Controls.Compatibility.Hosting;
 using Microsoft.Maui.Controls.Hosting;
 using Microsoft.Maui.Hosting;
+using Sample.Platforms.Tizen;
 using Tizen.Theme.Common;
 using Tizen.TV.UIControls.Forms;
 
@@ -15,6 +16,11 @@ namespace Sample
             builder
                 .UseMauiApp<App>()
                 .UseTizenTVUIControls()
+                .ConfigureMauiHandlers(handlers =>
+                {
+                    // This is a workaround for wrong ListView behavior
+                    handlers.AddHandler<ListView, TVListViewRenderer>();
+                })
                 .ConfigureFonts(fonts =>
                 {
                     fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
